Fixes vars types

+++ Project 2/sieve_sms.cpp
@@ -44,7 +44,7 @@ int main (int argc, char *argv[])
     
     // PAPI 
 	int EventSet = PAPI_NULL;
-  	long long values[2];
+  	ll values[2];
   	int ret;
 
 	ret = PAPI_library_init( PAPI_VER_CURRENT );
